[OBEY-CAMPAIGN-8deed8b4] refactor: split oversized handlers.go and navigator.go (refs #206 item 1)
Both files exceeded the 500-line guideline. Pure file reorganization within
each package; no behavior change (functions moved verbatim, imports pruned by
goimports).

- internal/commands/status/handlers.go (693 -> 393): listing handlers
  (festival/phase/sequence/task listing) moved to handlers_listing.go.
- internal/guidance/workflow/navigator.go (749 -> 395): step transitions
  (MarkComplete/MarkSkipped/SkipCurrentStep/MarkFailed/Advance/GetProgress)
  moved to navigator_transitions.go; instruction/progress formatting moved to
  navigator_format.go.

All 26 status handlers and 33 Navigator methods preserved exactly (verified by
function count). just lint all 0 issues; just test unit-raw 91 packages pass.
